preserve only new RPT client
