+++ .pipeline/checkstyle-suppressions.xml
@@ -10,7 +10,6 @@
   <suppress files="[/\\]core[/\\]model[/\\]" checks=".*"/>
   <suppress files="[/\\]openai[/\\]generated[/\\]model[/\\]" checks=".*"/>
   <suppress files="[/\\]rpt[/\\]generated[/\\]" checks=".*"/>
-  <suppress files="[/\\]rpt1_5[/\\]generated[/\\]" checks=".*"/>
   <suppress files="[/\\]orchestration[/\\]model[/\\]" checks=".*"/>
   <suppress files="[/\\]grounding[/\\]client[/\\]" checks=".*"/>
   <suppress files="[/\\]grounding[/\\]model[/\\]" checks=".*"/>
