+++ .pipeline/spotbugs-exclusions.xml
@@ -8,7 +8,7 @@
       <Package name="com.sap.ai.sdk.grounding.client"/>
       <Package name="com.sap.ai.sdk.grounding.model"/>
       <Package name="com.sap.ai.sdk.foundationmodels.openai.generated.model"/>
-      <Package name="com.sap.ai.sdk.foundationmodels.rpt1_5.generated.model"/>
+      <Package name="com.sap.ai.sdk.foundationmodels.rpt.generated.model"/>
       <Package name="com.sap.ai.sdk.prompt.registry.model"/>
       <Package name="com.sap.ai.sdk.prompt.registry.client"/>
     </Or>
