+++ foundation-models/sap-rpt/src/main/java/com/sap/ai/sdk/foundationmodels/rpt/RptClient.java
@@ -2,7 +2,6 @@ package com.sap.ai.sdk.foundationmodels.rpt;
 
 import static com.sap.ai.sdk.core.JacksonConfiguration.getDefaultObjectMapper;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.common.annotations.Beta;
 import com.sap.ai.sdk.core.AiCoreService;
 import com.sap.ai.sdk.core.DeploymentResolutionException;
@@ -14,6 +13,7 @@ import com.sap.ai.sdk.foundationmodels.rpt.generated.model.PredictionConfig;
 import com.sap.cloud.sdk.cloudplatform.connectivity.Destination;
 import com.sap.cloud.sdk.services.openapi.apache.apiclient.ApiClient;
 import java.io.File;
+import java.io.IOException;
 import java.util.Map;
 import javax.annotation.Nonnull;
 import lombok.AccessLevel;
@@ -107,10 +107,11 @@ public class RptClient {
   public PredictResponsePayload tableCompletion(
       @Nonnull final File parquetFile, @Nonnull final PredictionConfig predictionConfig) {
     try {
-      final var config =
-          JacksonConfiguration.getDefaultObjectMapper().writeValueAsString(predictionConfig);
-      return api.predictParquet(parquetFile, config);
-    } catch (JsonProcessingException e) {
+      final File tempFile = File.createTempFile("prediction-config", ".json");
+      tempFile.deleteOnExit();
+      JacksonConfiguration.getDefaultObjectMapper().writeValue(tempFile, predictionConfig);
+      return api.predictParquet(parquetFile, tempFile);
+    } catch (IOException e) {
       throw new IllegalArgumentException("Failed to serialize PredictionConfig to JSON", e);
     }
   }
