--- foundation-models/sap-rpt/src/main/java/com/sap/ai/sdk/foundationmodels/rpt1_5/RptClient.java
@@ -1,118 +1,0 @@
-package com.sap.ai.sdk.foundationmodels.rpt1_5;
-
-import static com.sap.ai.sdk.core.JacksonConfiguration.getDefaultObjectMapper;
-
-import com.google.common.annotations.Beta;
-import com.sap.ai.sdk.core.AiCoreService;
-import com.sap.ai.sdk.core.DeploymentResolutionException;
-import com.sap.ai.sdk.core.JacksonConfiguration;
-import com.sap.ai.sdk.foundationmodels.rpt1_5.generated.client.DefaultApi;
-import com.sap.ai.sdk.foundationmodels.rpt1_5.generated.model.PredictRequestPayload;
-import com.sap.ai.sdk.foundationmodels.rpt1_5.generated.model.PredictResponsePayload;
-import com.sap.ai.sdk.foundationmodels.rpt1_5.generated.model.PredictionConfig;
-import com.sap.cloud.sdk.cloudplatform.connectivity.Destination;
-import com.sap.cloud.sdk.services.openapi.apache.apiclient.ApiClient;
-import java.io.File;
-import java.io.IOException;
-import java.util.Map;
-import javax.annotation.Nonnull;
-import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
-
-/**
- * Client for interacting with SAP RPT foundation models.
- *
- * @since 1.16.0
- */
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public class RptClient {
-  @Nonnull private final DefaultApi api;
-  @Nonnull private final DefaultApi apiWithGzipEncoding;
-
-  /**
-   * Creates a new RptClient for the specified foundation model.
-   *
-   * @param foundationModel The foundation model to use.
-   * @return A new instance of RptClient.
-   * @throws DeploymentResolutionException If there is an error resolving the deployment.
-   */
-  @Nonnull
-  public static RptClient forModel(@Nonnull final RptModel foundationModel)
-      throws DeploymentResolutionException {
-    final var destination = new AiCoreService().getInferenceDestination().forModel(foundationModel);
-    return forDestination(destination);
-  }
-
-  /**
-   * Creates a new RptClient for the specified destination.
-   *
-   * @param destination The destination to use.
-   * @return A new instance of RptClient.
-   */
-  static RptClient forDestination(@Nonnull final Destination destination) {
-    final var apiClient = ApiClient.create(destination).withObjectMapper(getDefaultObjectMapper());
-    final var api = new DefaultApi(apiClient);
-    return new RptClient(api, api.withDefaultHeaders(Map.of("Content-Encoding", "gzip")));
-  }
-
-  /**
-   * Predict targets using SAP RPT model with structured data.
-   *
-   * <p>Note: This method is marked as {@link Beta} because it uses generated API types in its
-   * public signature.
-   *
-   * <p><b>200</b> - Successful Prediction
-   *
-   * <p><b>400</b> - Bad Request - Invalid input data
-   *
-   * <p><b>413</b> - Payload Too Large
-   *
-   * <p><b>422</b> - Validation Error
-   *
-   * <p><b>500</b> - Internal Server Error
-   *
-   * @param requestBody The prediction request
-   * @return prediction response from the RPT model
-   */
-  @Beta
-  @Nonnull
-  public PredictResponsePayload tableCompletion(@Nonnull final PredictRequestPayload requestBody) {
-    return apiWithGzipEncoding.predict(requestBody);
-  }
-
-  /**
-   * Make in-context predictions for specified target columns based on provided table data Parquet
-   * file.
-   *
-   * <p>Note: This method is marked as {@link Beta} because it uses generated API types in its
-   * public signature.
-   *
-   * <p><b>200</b> - Successful Prediction
-   *
-   * <p><b>400</b> - Bad Request - Invalid input data
-   *
-   * <p><b>413</b> - Payload Too Large
-   *
-   * <p><b>422</b> - Validation Error
-   *
-   * <p><b>500</b> - Internal Server Error
-   *
-   * @param parquetFile Parquet file
-   * @param predictionConfig The prediction configuration
-   * @return prediction response from the RPT model
-   * @since 1.16.0
-   */
-  @Beta
-  @Nonnull
-  public PredictResponsePayload tableCompletion(
-      @Nonnull final File parquetFile, @Nonnull final PredictionConfig predictionConfig) {
-    try {
-      final File tempFile = File.createTempFile("prediction-config", ".json");
-      tempFile.deleteOnExit();
-      JacksonConfiguration.getDefaultObjectMapper().writeValue(tempFile, predictionConfig);
-      return api.predictParquet(parquetFile, tempFile);
-    } catch (IOException e) {
-      throw new IllegalArgumentException("Failed to serialize PredictionConfig to JSON", e);
-    }
-  }
-}
