+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/controllers/RptController.java
@@ -1,7 +1,7 @@
 package com.sap.ai.sdk.app.controllers;
 
 import com.sap.ai.sdk.app.services.RptService;
-import com.sap.ai.sdk.foundationmodels.rpt1_5.generated.model.PredictionsInnerValue;
+import com.sap.ai.sdk.foundationmodels.rpt.generated.model.PredictionsInnerValue;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
